+++ src/mltraq/storage/serialization.py
@@ -1,6 +1,6 @@
 import datetime
 import uuid
-from typing import List, Tuple
+from typing import Any, List, Tuple
 
 import pandas as pd
 from numpy import float32, float64, int32, int64
@@ -45,7 +45,7 @@ def serialize(obj: object) -> bytes:
     return serializer.serialize(obj)
 
 
-def deserialize(data: bytes) -> object:
+def deserialize(data: bytes) -> Any:
     """
     Deserialize object, using the preferred serializer.
     """
@@ -70,7 +70,7 @@ def unsafe_pickle(obj: object) -> bytes:
     return PickleSerializer.serialize(obj, assert_safe=False)
 
 
-def unsafe_unpickle(data: bytes) -> object:
+def unsafe_unpickle(data: bytes) -> Any:
     """
     Unpickle object, without limiting to safe opcodes.
     Used to unpickle complete Experiment objects.