+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Recent changes
 
+## 0.1.153
+* Added Python and MLtraq version to experiment's metadata
+* Added `utils/sysenv.py` to track environment stats (system, architecture, locale, time, ...)
+* Added `utils/reproducibility.py`, providing a temporary seed context
+* Added `utils/plotting.py`, providing a temporary plot context for a single axes
+* Removed `utils/plot.py` due to lack of usage (and unlikely usage in the future)
+* Improved `BaseOptions.get(...)` with `prefer` and `otherwise` parameters to make it more flexible
+
 ## 0.1.152
 * Increased tests to `87%` coverage